today git commit

+++ database/migrations/2022_06_12_095106_create_price_lists_table.php
@@ -15,7 +15,7 @@ class CreatePriceListsTable extends Migration
     {
         Schema::create('price_lists', function (Blueprint $table) {
             $table->id();
-            $table->tinyInteger('price_sl')->unsigned();
+            $table->Integer('price_sl')->unsigned();
             $table->decimal('amount');
             $table->softDeletes();
             $table->timestamps();
